reorganize navigations
